dnsdist: Add steal, iowait and UDP errors metrics
This commit also moves the prometheus metrics code to a separate
header.
